services/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Task scheduling
apscheduler==3.10.4

# LangGraph for agent orchestration
langgraph==0.0.62
langchain-core>=0.2.0

# Data validation
email-validator==2.1.0

# Web scraping and parsing
beautifulsoup4==4.12.2
selenium==4.15.2
google-search-results==2.4.2  # SerpAPI
lxml==4.9.3

# AI/ML and language processing
openai==1.3.5
langchain==0.0.340
langchain-openai==0.0.2
scikit-learn==1.3.2

# Document processing
python-docx==0.8.11
PyPDF2==3.0.1
pypdf==3.17.1

# Utilities and retry logic
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
